split extension command constant on new file
