+++ src/extension/src/client-modules/generation.ts
@@ -1,7 +1,6 @@
 import * as vscode from "vscode";
 import { WizardServant, IPayloadResponse } from "../wizardServant";
 import {
-  ExtensionCommand,
   TelemetryEventName,
   CONSTANTS,
   AzureResourceType
@@ -15,6 +14,7 @@ import { Logger } from "../utils/logger";
 import { IGenerationPayloadType, IServicesGenerationPayload } from "../types/generationPayloadType";
 import { sendToClientGenerationStatus, GenerationItemStatus, updateStatusMessage, sendToClientGenerationPath, GENERATION_NAMES } from "../utils/generationStatus";
 import { MESSAGES } from "../constants/messages";
+import { ExtensionCommand } from "../constants/extension";
 
 interface DeployedServiceStatus {
   serviceType: AzureResourceType;
