+++ src/extension/src/utils/defaults.ts
@@ -1,9 +1,10 @@
 import { WizardServant, IPayloadResponse } from "../wizardServant";
-import { ExtensionCommand, PROJECT_NAME_VALIDATION_LIMIT } from "../constants/constants";
+import { PROJECT_NAME_VALIDATION_LIMIT } from "../constants/constants";
 import * as vscode from "vscode";
 import * as os from "os";
 import * as fs from "fs";
 import * as path from "path";
+import { ExtensionCommand } from "../constants/extension";
 
 export class Defaults extends WizardServant {
   clientCommandMap: Map<ExtensionCommand,(message: any) => Promise<IPayloadResponse>>;
